If label is not set for a field, automatically convert the field name to a title
We'll convert camel case to a lowercase label where words are separate by space.
If there is an uppercase abbreviation we'll leave it untouched.
Underscore chars will be replaced with spaces.
If the field name contains a number than we'll add a space before the number.

e.g. "myField" => "my field", "myCMSId" => "my CMS id"
     "my_field" => "my field", "myField1" => "my field 1"